chore: update vite dev and build process
