--- entry-popup.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta name="color-scheme" content="dark">
-        <title>Userscripts popup</title>
-    </head>
-    <body>
-        <noscript>Please enable Javascript to use this app</noscript>
-        <div id="app"></div>
-        <script type="module" src="/src/popup/main.js"></script>
-    </body>
-</html>
