--- index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Userscripts dev</title>
-        <style>
-            :root {
-                color-scheme: light dark;
-                background-color: #242424;
-            }
-        </style>
-    </head>
-    <body>
-        <div id="app"></div>
-        <script type="module" src="/src/main.js"></script>
-        <script>
-            function notice() {
-                if (!app.hasChildNodes()) {
-                    app.append('Please visit in the Userscripts development environment.');
-                };
-            };
-            setTimeout(notice, 500);
-        </script>
-    </body>
-</html>
