+++ src/page/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta name="color-scheme" content="dark" />
+		<title>Userscripts page</title>
+	</head>
+	<body>
+		<noscript>Please enable Javascript to use this app</noscript>
+		<div id="app"></div>
+		<script type="module" src="./main.js"></script>
+	</body>
+</html>
