+++ xcode/Safari-Extension/Resources/manifest.json
@@ -17,7 +17,7 @@
         "persistent": false
     },
     "browser_action": {
-        "default_popup": "dist/entry-popup.html",
+        "default_popup": "popup.html",
         "default_icon": {
             "16": "images/toolbar-icon-16.png",
             "32": "images/toolbar-icon-32.png"
